requirements.txt
customtkinter
pyserial
pynput